requirements.txt
httpx>=0.28.1
pydantic==2.11.9
pydantic_ai==1.0.12
pydantic_settings==2.11.0
python-dotenv==1.1.1
streamlit==1.46.1
openai>=1.56.0
pandas>=2.0.0
python-dateutil>=2.8.0
streamlit-aggrid>=1.0.5
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
llama-index-core>=0.11.0
llama-index-embeddings-openai>=0.2.0
llama-index-llms-openai>=0.2.0
selenium>=4.0.0
boto3>=1.26.0
botocore>=1.34.0
streamlit-agraph>=0.0.43
llama-index-embeddings-azure-openai
streamlit
streamlit-agraph
streamlit-aggrid
anthropic
pydantic
pydantic-ai
pydantic-ai-slim[openai]
pandas
numpy
requests
beautifulsoup4
lxml
selenium
webdriver-manager
boto3
python-dotenv
langchain
langchain-community
langchain-openai
llama-index
llama-index-embeddings-openai
llama-index-embeddings-azure-openai
llama-index-llms-openai
llama-index-llms-azure-openai
openai
plotly
Pillow
openpyxl
gensim>=4.0.0
